requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.13.0
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.2
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
arize-phoenix==12.19.0
arize-phoenix-client==1.26.0
arize-phoenix-evals==2.7.0
arize-phoenix-otel==0.14.0
astor==0.8.1
asyncio==4.0.0
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
blake3==1.0.8
blinker==1.9.0
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
compressed-tensors==0.12.2
cryptography==46.0.3
dataclasses-json==0.6.7
datasets==4.4.1
deprecation==2.1.0
depyf==0.20.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.6.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.72.0
gotrue==2.12.4
graphql-core==3.2.7
greenlet==3.3.0
grpc-interceptor==0.15.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
httpx-sse==0.4.3
httpx-ws==0.8.2
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.8
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langsmith==0.4.46
lark==1.2.2
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
lmstudio==1.5.0
loguru==0.7.3
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mistral_common==1.8.5
mlx==0.30.0
mlx-lm==0.28.3
mlx-metal==0.30.0
model-hosting-container-standards==0.1.9
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.12.0
networkx==3.6
ninja==1.13.0
numba==0.61.2
numpy==2.3.5
ollama==0.6.1
openai==2.8.1
openai-harmony==0.0.8
opencv-python-headless==4.12.0.88
openinference-instrumentation==0.1.42
openinference-semantic-conventions==0.1.25
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
outlines_core==0.2.11
packaging==25.0
pandas==2.3.3
partial-json-parser==0.2.1.1.post7
pgvector==0.4.1
pillow==12.0.0
pluggy==1.6.0
ply==3.11
postgrest==0.16.11
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyarrow==21.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.4.0
pystache==0.6.8
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rank-bm25==0.2.2
realtime==1.0.6
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
rpds-py==0.29.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentencepiece==0.2.1
sentry-sdk==2.46.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlean.py==3.50.4.5
starlette==0.50.0
storage3==0.7.7
strawberry-graphql==0.270.1
streamlit==1.51.0
StrEnum==0.4.15
supabase==2.4.3
supabase-auth==2.24.0
supabase-functions==2.24.0
supafunc==0.4.7
supervisor==4.3.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.11.2
watchfiles==1.1.1
websockets==12.0
wrapt==1.17.3
wsproto==1.3.2
xgrammar==0.1.25
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
